Seperated game objects and main scene
