+++ src/game_objects/game_object.cpp
@@ -15,13 +15,7 @@ void GameObject::draw(Shader *shader, SceneInfo &scene_info) {
 }
 
 // Updates Game Object Paramters
-void GameObject::update() {
-    // Test code for playing animation for the boy
-    anim_player.set_speed(3.0f);
-
-    if (model->animations.size() > 0) {
-        anim_player.play(model, "open");        
-    }
+void GameObject::update() {    
 }
 
 void GameObject::set_color(glm::vec3 color) {
